Fix modified Utf8 encoding, add some linking validation
`names.wast` test case passes:

 * Fixed (hopefully for the last time) the modified UTF8 encoding.
   Added unit tests using `java.io.DataOutputStream#writeUTF` for the
   various edge cases that have been problematic in the past

 * Fixed the harness to properly escape string literals into the UTF-16
   format Java expects.

 * Disabled generating methods for function exports (since this test
   cases really stresses the renamer in a way that doesn't matter)

Linking validation is much improved:

 * Up to multiple-value return types, imported functions have their
   types validated

 * Imported globals have their types validated, though mutability checks
   are still missing

Added in several more of the missing `spectest` exports.

+++ lib/src/jvm/code/code_builder_exts.rs
@@ -300,7 +300,7 @@ impl<'g> CodeBuilderExts<'g> for CodeBuilder<'g> {
                 let type_static_field = match base_type {
                     BaseType::Int => self.java.members.lang.integer.r#type,
                     BaseType::Long => self.java.members.lang.long.r#type,
-                    BaseType::Float => self.java.members.lang.long.r#type,
+                    BaseType::Float => self.java.members.lang.float.r#type,
                     BaseType::Double => self.java.members.lang.double.r#type,
                     BaseType::Boolean => self.java.members.lang.boolean.r#type,
                     other => todo!("const_class for {:?}", other),
